requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.44.0
colorama==0.4.6
Django==5.2
django-cloudinary-storage==0.3.0
djangorestframework==3.16.0
flake8==7.2.0
idna==3.10
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pycodestyle==2.13.0
pyflakes==3.3.2
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
